requirements.txt
# Flask 框架
flask

# LINE Bot SDK
line-bot-sdk

# 環境變數管理
python-dotenv

# # Whisper 語音轉文字
# git+https://github.com/openai/whisper.git
# ffmpeg-python

# Together.ai API 模型支援
openai

# 測試與除錯工具
pytest
pytest-mock

# 開發工具
black
isort
flake8

# 其他依賴
requests
pandas
gunicorn

